refactored backend structure
