+++ dronuts-backend/.eslintrc.yml
@@ -3,14 +3,12 @@ env:
   commonjs: true
   es2021: true
 extends:
-  - plugin:react/recommended
   - google
 parserOptions:
   ecmaFeatures:
     jsx: true
   ecmaVersion: latest
-plugins:
-  - react
+plugins: []
 rules: 
   indent: 
     - error
