+++ dronuts-backend/src/drone_schema.js
